requirements.txt

fastapi
uvicorn[standard]
beautifulsoup4
requests
selenium
webdriver-manager
python-Wappalyzer
undetected-chromedriver
builtwith
supabase
python-dotenv

# --- New RAG Libraries ---
langchain
langchain-community
langchain-ollama
# FAISS is a local vector store from Facebook AI
faiss-cpu
# SentenceTransformers is for creating embeddings locally
sentence-transformers